refactor(activemodel): extract inspectAccessor to shared helper
Both acceptance.ts and confirmation.ts had divergent copies of the
same prototype-walk helper. Move it to validations/_accessor.ts and
import from both. Also tightens the setupBang docstring on
acceptance.ts to match actual skip semantics (skip only when both
sides exist; otherwise install the missing half).
