+++ packages/activemodel/src/validations/confirmation.ts
@@ -1,6 +1,7 @@
 import { EachValidator } from "../validator.js";
 import type { AnyRecord } from "../validator.js";
 import { humanize } from "@blazetrails/activesupport";
+import { inspectAccessor } from "./_accessor.js";
 
 /**
  * Mirrors: ActiveModel::Validations::ConfirmationValidator (confirmation.rb)
@@ -75,65 +76,6 @@ export function setupBang(this: ConfirmationHost, klass: unknown): void {
         },
     });
   }
-}
-
-interface InheritedAccessor {
-  hasGetter: boolean;
-  hasSetter: boolean;
-  getter?: (this: object) => unknown;
-  setter?: (this: object, value: unknown) => void;
-}
-
-/**
- * Walk the prototype chain and capture the first-found accessor
- * shape for `name`. Distinguishes accessor descriptors (`get`/`set`)
- * from data descriptors (`"value" in desc` — handles the
- * `value: undefined` case correctly) and synthesizes a getter/setter
- * pair from a data property so callers can preserve it in a new
- * accessor descriptor.
- */
-function inspectAccessor(prototype: object, name: string): InheritedAccessor {
-  let proto: object | null = prototype;
-  while (proto && proto !== Object.prototype) {
-    const desc = Object.getOwnPropertyDescriptor(proto, name);
-    if (desc) {
-      if ("value" in desc || "writable" in desc) {
-        // Plain data property — synthesize a getter/setter that reads
-        // through the captured prototype slot via Reflect.get, and
-        // writes via own-property defineProperty. Reading `this[name]`
-        // would recurse into the new accessor we're about to install.
-        const inheritedProto = proto;
-        const enumerable = desc.enumerable ?? true;
-        const configurable = desc.configurable ?? true;
-        return {
-          hasGetter: true,
-          hasSetter: desc.writable !== false,
-          getter() {
-            return Reflect.get(inheritedProto, name, this);
-          },
-          setter:
-            desc.writable !== false
-              ? function (this: object, v: unknown) {
-                  Object.defineProperty(this, name, {
-                    value: v,
-                    writable: true,
-                    enumerable,
-                    configurable,
-                  });
-                }
-              : undefined,
-        };
-      }
-      return {
-        hasGetter: typeof desc.get === "function",
-        hasSetter: typeof desc.set === "function",
-        getter: desc.get as ((this: object) => unknown) | undefined,
-        setter: desc.set as ((this: object, v: unknown) => void) | undefined,
-      };
-    }
-    proto = Object.getPrototypeOf(proto);
-  }
-  return { hasGetter: false, hasSetter: false };
 }
 
 /**
